+++ test/test_osrelease.py
@@ -1,12 +1,12 @@
 import os
 import unittest
 
-import osbuild
+from osbuild.util import osrelease
 
 class TestOSRelease(unittest.TestCase):
     def test_non_existant(self):
         """Verify default os-release value, if no files are given."""
-        self.assertEqual(osbuild.pipeline.describe_os(), "linux")
+        self.assertEqual(osrelease.describe_os(), "linux")
 
     def test_describe_os(self):
         """Test host os detection. test/os-release contains the os-release files
@@ -14,4 +14,4 @@ class TestOSRelease(unittest.TestCase):
         """
         for entry in os.scandir("test/os-release"):
             with self.subTest(entry.name):
-                self.assertEqual(osbuild.pipeline.describe_os(entry.path), entry.name)
+                self.assertEqual(osrelease.describe_os(entry.path), entry.name)
